+++ main.c
@@ -104,7 +104,7 @@ int get_int(const int min, const int max) {
 /**
  * Return the sum of a list of integers.
  */
-int sum_list_int(int *list, int list_len) {
+int sum_list_int(const int list[], const int list_len) {
     int sum = 0;
     for (int i = 0; i < list_len; i++) {
         sum += list[i];
@@ -115,12 +115,67 @@ int sum_list_int(int *list, int list_len) {
 /**
  * Return the sum of a list of floats.
  */
-float sum_list_float(float *list, int list_len) {
+float sum_list_float(const float list[], const int list_len) {
     float sum = 0;
     for (int i = 0; i < list_len; i++) {
         sum += list[i];
     }
     return sum;
+}
+
+void quicksort_recursive(int *coords, const int low, const int high, const int width) {
+
+    if (low < high) {
+
+        /*
+        Pivot value is the value at coords[high]. Every value less than pivot
+        will be to the left of wherever pivot ends up. Remember: a coord is
+        {x, y, index (of equivalent dot in dots)}. The array isn't 2D as it
+        requires malloc.
+        */
+        int pivot = coords[high * 3 + 1] * width + coords[high * 3];
+
+        // Move Smaller Elements to the Left
+
+        int i = low - 1;
+
+        for (int ii = low; ii <= high - 1; ii++) {
+            if (coords[ii * 3 + 1] * width + coords[ii * 3] < pivot) {
+                i++;
+                const int temp[3] = {coords[i * 3], coords[i * 3 + 1], coords[i * 3 + 2]};
+                coords[i * 3] = coords[ii * 3];
+                coords[i * 3 + 1] = coords[ii * 3 + 1];
+                coords[i * 3 + 2] = coords[ii * 3 + 2];
+                coords[ii * 3] = temp[0];
+                coords[ii * 3 + 1] = temp[1];
+                coords[ii * 3 + 2] = temp[2];
+            }
+        }
+
+        // Move Pivot to After Smaller Elements
+
+        i++;
+        const int temp[3] = {coords[i * 3], coords[i * 3 + 1], coords[i * 3 + 2]};
+        coords[i * 3] = coords[high * 3];
+        coords[i * 3 + 1] = coords[high * 3 + 1];
+        coords[i * 3 + 2] = coords[high * 3 + 2];
+        coords[high * 3] = temp[0];
+        coords[high * 3 + 1] = temp[1];
+        coords[high * 3 + 2] = temp[2];
+
+        // Decide Whether Recursion is Needed
+        /*
+        Cases:
+        i == med_index, the median is in the correct spot, no more sorting
+        i > med_index, median is in left section, left section needs sorting
+        i < med_index, median in right section, sort right section
+        */
+
+        quicksort_recursive(coords, low, i - 1, width);
+        quicksort_recursive(coords, i + 1, high, width);
+
+    }
+
 }
 
 
@@ -134,8 +189,8 @@ float sum_list_float(float *list, int list_len) {
  * based on. HIGH and LOW determine the sorting bounds for a given level of
  * recursion.
  */
-void nth_sort_recursive(
-    int coords[], const int low, const int high, const int axis, const int med_index
+void median_sort_recursive(
+    int *coords, const int low, const int high, const int axis, const int med_index
 ) {
 
     if (low < high) {
@@ -185,9 +240,9 @@ void nth_sort_recursive(
         */
 
         if (i > med_index) {
-            nth_sort_recursive(coords, low, i - 1, axis, med_index);
+            median_sort_recursive(coords, low, i - 1, axis, med_index);
         } else if (i < med_index) {
-            nth_sort_recursive(coords, i + 1, high, axis, med_index);
+            median_sort_recursive(coords, i + 1, high, axis, med_index);
         }
 
     }
@@ -198,9 +253,10 @@ void nth_sort_recursive(
  * Build a KDTree from COORDS, and return the root node. Ensures the KDTree is
  * built with the lowest possible depth for maximum efficiency when querying the
  * tree. Every recursion creates one dot and calls this function to insert its
- * children from an array of possible dots. DEPTH should be 0.
+ * children from an array of possible dots. DEPTH should be 0. COORDS should be
+ * of length NUM_COORDS * 3.
  */
-Node *build_recursive(const int num_coords, int coords[num_coords * 3], const int depth) {
+Node *build_recursive(int *coords, const int num_coords, const int depth) {
 
     const int med_pos = num_coords / 2;
 
@@ -210,7 +266,7 @@ Node *build_recursive(const int num_coords, int coords[num_coords * 3], const in
     everything else to the right.
     */
 
-    nth_sort_recursive(coords, 0, num_coords - 1, depth % 2, med_pos);
+    median_sort_recursive(coords, 0, num_coords - 1, depth % 2, med_pos);
 
     // Add Median Node to Tree
     // Every recursion adds one node
@@ -234,7 +290,7 @@ Node *build_recursive(const int num_coords, int coords[num_coords * 3], const in
             coords_left[i * 3 + 1] = coords[i * 3 + 1];
             coords_left[i * 3 + 2] = coords[i * 3 + 2];
         }
-        node->left = build_recursive(num_coords_left, coords_left, depth + 1);
+        node->left = build_recursive(coords_left, num_coords_left, depth + 1);
         free(coords_left);
 
         // Decide whether node will have a right child
@@ -249,7 +305,7 @@ Node *build_recursive(const int num_coords, int coords[num_coords * 3], const in
                 coords_right[i * 3 + 1] = coords[index * 3 + 1];
                 coords_right[i * 3 + 2] = coords[index * 3 + 2];
             }
-            node->right = build_recursive(num_coords_right, coords_right, depth + 1);
+            node->right = build_recursive(coords_right, num_coords_right, depth + 1);
             free(coords_right);
         }
 
@@ -268,7 +324,7 @@ Node *build_recursive(const int num_coords, int coords[num_coords * 3], const in
  * node.
  */
 void query_recursive(
-    Node *node, const int *coord, const int depth, int *index_ptr, int *min_dist_ptr
+    Node *node, const int coord[2], const int depth, int *index_ptr, int *min_dist_ptr
 ) {
 
     // Calculate Distance
@@ -281,10 +337,12 @@ void query_recursive(
     // Update Minimum Distance and Index Pointers
 
     if (dist < *min_dist_ptr) {
+        *min_dist_ptr = dist;
         if (index_ptr != NULL) {
             *index_ptr = node->index;
+        } else if (dist < 15 * 15) {
+            return; // Shortcut specifically for water biome generation
         }
-        *min_dist_ptr = dist;
     }
 
     // Decide Whether Recursion is Needed
@@ -315,8 +373,8 @@ void query_recursive(
  * DEPTH should be 0 when NODE is a root node.
  */
 void query_dist_recursive(
-    Node *node, const int *coord, const int depth,
-    int *dists, const int dists_len
+    Node *node, const int coord[2], const int depth,
+    int dists[], const int dists_len
 ) {
 
     // Calculate Distance
@@ -385,7 +443,7 @@ void free_recursive(Node *node) {
  * as well. The length of TYPE + NUM (as a string) must be at max 9 characters
  * (including null character).
  */
-void set_process_title(const char *type, const int num) {
+void set_process_title(const char type[], const int num) {
 
     // Example: type is "worker", num is 6: "biogen-worker6"
 
@@ -427,8 +485,8 @@ void track_progress(
         "Setup", "Section Generation", "Section Assignment", "Coastline Smoothing",
         "Biome Generation", "Image Generation", "Finish"
     };
-    float section_weights[7] = {0.01, 0.01, 0.02, 0.60, 0.06, 0.10, 0.20};
-    // Used for overall progress bar (e.g. Setup takes ~1% of total time)
+    float section_weights[7] = {0.03, 0.01, 0.01, 0.14, 0.04, 0.24, 0.53};
+    // Used for overall progress bar (e.g. Setup takes ~3% of total time)
 
     while (true) {
 
@@ -455,13 +513,13 @@ void track_progress(
 
             // Check if Section Complete
 
-            char *color;
+            char color[10];
             float section_time;
             if (section_times[i] != 0.0) {
-                color = ANSI_GREEN;
+                strncpy(color, ANSI_GREEN, 10);
                 section_time = section_times[i]; // Section complete
             } else {
-                color = ANSI_BLUE;
+                strncpy(color, ANSI_BLUE, 10);
                 if (i == 0 || section_times[i - 1] != 0.0) {
                     section_time = time_diff - sum_list_float(section_times, 7);
                     // Section in progress
@@ -494,13 +552,13 @@ void track_progress(
 
         // Total Progress
 
-        char *color;
+        char color[10];
         float total_time;
         if (section_times[7] != 0.0) {
-            color = ANSI_GREEN;
+            strncpy(color, ANSI_GREEN, 10);
             total_time = section_times[7];
         } else {
-            color = ANSI_BLUE;
+            strncpy(color, ANSI_BLUE, 10);
             total_time = (float)(time_now.tv_sec - start_time.tv_sec) +
                 (time_now.tv_nsec - start_time.tv_nsec) / 1000000000.0;
         }
@@ -540,23 +598,32 @@ void track_progress(
  * dot's distance from the nearest land origin dot.
  */
 void assign_sections(
-    const int map_resolution, const float island_size, const int start_index, const int end_index,
+    const int map_resolution, const float island_size,
+    const int start_index, const int end_index, const int *reg_dots,
     Node *origin_tree_root, Dot *dots, _Atomic int *section_progress
 ) {
 
     srand(time(NULL) + getpid());
 
+    int min_dist;
+
     for (int i = start_index; i < end_index; i++) {
     // Non-water dots are not included
 
-        Dot *dot = &dots[i];
+        // Calculate Maximum Distance
+
+        if (i != start_index && reg_dots[i * 3 + 1] == reg_dots[(i - 1) * 3 + 1]) {
+            int min_dist_sq = (int)sqrt(min_dist) + 1 + reg_dots[i * 3] - reg_dots[(i - 1) * 3];
+            min_dist = min_dist_sq * min_dist_sq;
+        } else {
+            min_dist = INT_MAX;
+        }
 
         // Find Distance to Nearest Origin Dot
 
-        int min_dist = INT_MAX;
         // min and dist are squared, sqrt is not done until later
         int min_index = 0;
-        int coord[2] = {dot->x, dot->y};
+        int coord[2] = {reg_dots[i * 3], reg_dots[i * 3 + 1]};
         query_recursive(origin_tree_root, coord, 0, &min_index, &min_dist);
 
         // Calculate Chance
@@ -567,7 +634,7 @@ void assign_sections(
         int chance = (dist <= threshold) ? 9 : 1;
 
         if (rand() % 10 < chance) {
-            dot->type = 'L'; // Land
+            dots[reg_dots[i * 3 + 2]].type = 'L'; // Land
         }
 
         atomic_fetch_add(&section_progress[2], 1);
@@ -582,62 +649,125 @@ void assign_sections(
  * COASTLINE_SMOOTHING dots of the same and opposite types.
  */
 void smooth_coastlines(
-    const int coastline_smoothing, const int start_index, const int end_index,
-    Node *land_tree_root, Node *water_tree_root,
-    const int num_dots, const int num_special_dots, const int num_reg_dots,
+    const int coastline_smoothing,
+    const int *land_dots, const int land_start, const int land_end, Node *land_tree_root,
+    const int *water_dots, const int water_start, const int water_end, Node *water_tree_root,
+    const int num_dots, const int num_land_dots, const int num_water_dots,
     Dot *dots, _Atomic int *section_progress
 ) {
 
-    // Smooth Coastlines
+    int dists_same[coastline_smoothing];
+    int dists_opp[coastline_smoothing];
 
-    for (int i = start_index; i < end_index; i++) {
-    // Only includes "Land" and "Water" dots
+    // Coastline Smoothing for Land Dots
 
-        // Get Dot Information
+    for (int i = land_start; i < land_end; i++) {
 
-        Dot *dot = &dots[i];
-        int dot_coord[2] = {dot->x, dot->y};
+        int dot_coord[2] = {land_dots[i * 3], land_dots[i * 3 + 1]};
 
-        const bool land_dot = (dot->type == 'L');
+        // Calculate Maximum Distances
 
-        // Get Nearest Coastline_Smoothing Dots of Each Type
-
-        int dists_same[coastline_smoothing];
-        int dists_opp[coastline_smoothing];
-        for (int i = 0; i < coastline_smoothing; i++) {
-            dists_same[i] = INT_MAX;
+        bool same_y = (i != land_start && land_dots[i * 3 + 1] == land_dots[(i - 1) * 3 + 1]);
+        if (same_y) {
+            const int prev_dot_dist = land_dots[i * 3] - land_dots[(i - 1) * 3];
+            int min_dist_same = (int)sqrt(dists_same[coastline_smoothing - 1]) + 1 + prev_dot_dist;
+            // + 1 needed for floating-point errors (ceil didn't work)
+            min_dist_same *= min_dist_same;
+            int min_dist_opp = (int)sqrt(dists_opp[coastline_smoothing - 1]) + 1 + prev_dot_dist;
+            min_dist_opp *= min_dist_opp;
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_same[ii] = min_dist_same;
+                dists_opp[ii] = min_dist_opp;
+            }
+        } else {
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_same[ii] = INT_MAX;
+            }
         }
+
+        // Get Nearest Distances for Each Type
+
         long sum_same = 0;
         long sum_opp = 0;
-        Node *root_same;
-        Node *root_opp;
 
-        if (land_dot) {
-            root_same = land_tree_root;
-            root_opp = water_tree_root;
-        } else {
-            root_same = water_tree_root;
-            root_opp = land_tree_root;
+        query_dist_recursive(land_tree_root, dot_coord, 0, dists_same, coastline_smoothing);
+        for (int ii = 0; ii < coastline_smoothing; ii++) {
+            sum_same += dists_same[ii];
         }
 
-        query_dist_recursive(root_same, dot_coord, 0, dists_same, coastline_smoothing);
-        for (int i = 0; i < coastline_smoothing; i++) {
-            sum_same += dists_same[i];
+        if (!same_y) {
+            const int max = (sum_same < INT_MAX) ? sum_same : INT_MAX;
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_opp[ii] = max;
+            }
         }
-        const int max = (sum_same < INT_MAX) ? sum_same : INT_MAX;
-        for (int i = 0; i < coastline_smoothing; i++) {
-            dists_opp[i] = max;
-        }
-        query_dist_recursive(root_opp, dot_coord, 0, dists_opp, coastline_smoothing);
 
-        for (int i = 0; i < coastline_smoothing; i++) {
-            sum_opp += dists_opp[i];
+        query_dist_recursive(water_tree_root, dot_coord, 0, dists_opp, coastline_smoothing);
+        for (int ii = 0; ii < coastline_smoothing; ii++) {
+            sum_opp += dists_opp[ii];
         }
 
         // Change Dot Type if Closer to Opposite Type
 
         if (sum_same > sum_opp) {
-            dot->type = (land_dot) ? 'W' : 'L';
+            dots[land_dots[i * 3 + 2]].type = 'W';
+        }
+
+        atomic_fetch_add(&section_progress[3], 1);
+
+    }
+
+    // Coastline Smoothing for Water Dots
+
+    for (int i = water_start; i < water_end; i++) {
+
+        int dot_coord[2] = {water_dots[i * 3], water_dots[i * 3 + 1]};
+
+        // Calculate Maximum Distances
+
+        bool same_y = (i != water_start && water_dots[i * 3 + 1] == water_dots[(i - 1) * 3 + 1]);
+        if (same_y) {
+            const int prev_dot_dist = water_dots[i * 3] - water_dots[(i - 1) * 3];
+            int min_dist_same = (int)sqrt(dists_same[coastline_smoothing - 1]) + 1 + prev_dot_dist;
+            min_dist_same *= min_dist_same;
+            int min_dist_opp = (int)sqrt(dists_opp[coastline_smoothing - 1]) + 1 + prev_dot_dist;
+            min_dist_opp *= min_dist_opp;
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_same[ii] = min_dist_same;
+                dists_opp[ii] = min_dist_opp;
+            }
+        } else {
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_same[ii] = INT_MAX;
+            }
+        }
+
+        // Get Nearest Distances for Each Type
+
+        long sum_same = 0;
+        long sum_opp = 0;
+
+        query_dist_recursive(water_tree_root, dot_coord, 0, dists_same, coastline_smoothing);
+        for (int ii = 0; ii < coastline_smoothing; ii++) {
+            sum_same += dists_same[ii];
+        }
+
+        if (!same_y) {
+            const int max = (sum_same < INT_MAX) ? sum_same : INT_MAX;
+            for (int ii = 0; ii < coastline_smoothing; ii++) {
+                dists_opp[ii] = max;
+            }
+        }
+
+        query_dist_recursive(land_tree_root, dot_coord, 0, dists_opp, coastline_smoothing);
+        for (int ii = 0; ii < coastline_smoothing; ii++) {
+            sum_opp += dists_opp[ii];
+        }
+
+        // Change Dot Type if Closer to Opposite Type
+
+        if (sum_same > sum_opp) {
+            dots[water_dots[i * 3 + 2]].type = 'L';
         }
 
         atomic_fetch_add(&section_progress[3], 1);
@@ -652,45 +782,50 @@ void smooth_coastlines(
  * distance to the nearest land dot.
  */
 void generate_biomes_water(
-    const int start_index, const int end_index, const int height, Node *land_tree_root,
-    const int num_dots, Dot *dots, _Atomic int *section_progress
+    const int start_index, const int end_index, const int *water_dots, Node *land_tree_root,
+    const int height, const int num_dots, Dot *dots, _Atomic int *section_progress
 ) {
 
-    // Generate Water Biomes
+    int land_dist;
 
     for (int i = start_index; i < end_index; i++) {
 
-        // Get Dot Info
-
-        Dot *dot = &dots[i];
-
-        if (dot->type != 'W') {
-            continue;
-        }
-
         // Calculate Distance to Equator
 
-        const float equator_dist = fabs((float)dot->y - height / 2.0) / height * 20.0;
+        const float equator_dist =
+            fabs((float)water_dots[i * 3 + 1] - height / 2.0) / height * 20.0;
+
+        // Calculate Maximum Land Distance
+
+        if (i != start_index && water_dots[i * 3 + 1] == water_dots[(i - 1) * 3 + 1]) {
+            int min_dist_sq =
+                (int)sqrt(land_dist) + 1 + water_dots[i * 3] - water_dots[(i - 1) * 3];
+            land_dist = min_dist_sq * min_dist_sq;
+        } else {
+            land_dist = INT_MAX;
+        }
 
         // Calculate Distance to Land
 
-        int land_dist_sq = INT_MAX;
-        const int coord[2] = {dot->x, dot->y};
-        query_recursive(land_tree_root, coord, 0, NULL, &land_dist_sq);
+        const int coord[2] = {water_dots[i * 3], water_dots[i * 3 + 1]};
+        query_recursive(land_tree_root, coord, 0, NULL, &land_dist);
 
         // Set Water Biome
 
+        char dot_type = 'W';
         if ( // Remember: all land distances are squared for efficiency
-            (land_dist_sq < 35 * 35 && equator_dist > 9) ||
-            (land_dist_sq < 25 * 25 && equator_dist > 8) ||
-            (land_dist_sq < 15 * 15 && equator_dist > 7)
+            (land_dist < 35 * 35 && equator_dist > 9) ||
+            (land_dist < 25 * 25 && equator_dist > 8) ||
+            (land_dist < 15 * 15 && equator_dist > 7)
         ) {
-            dot->type = 'I';
-        } else if (land_dist_sq < 18 * 18) {
-            dot->type = 's';
-        } else if (land_dist_sq >= 35 * 35) {
-            dot->type = 'd';
+            dot_type = 'I';
+        } else if (land_dist < 18 * 18) {
+            dot_type = 's';
+        } else if (land_dist >= 35 * 35) {
+            dot_type = 'd';
         }
+
+        dots[water_dots[i * 3 + 2]].type = dot_type;
 
         atomic_fetch_add(&section_progress[4], 1);
 
@@ -704,30 +839,33 @@ void generate_biomes_water(
  * dots whose biomes are already before this function.
  */
 void generate_biomes_land(
-    const int start_index, const int end_index, Node *origin_tree_root, const int num_dots,
-    const int *biome_origin_indexes, Dot *dots, _Atomic int *section_progress
+    const int start_index, const int end_index, const int *land_dots,
+    Node *origin_tree_root, const int biome_origin_indexes[],
+    const int num_dots, Dot *dots, _Atomic int *section_progress
 ) {
 
-    // Generate Land Biomes
+    int min_dist;
 
     for (int i = start_index; i < end_index; i++) {
 
-        Dot *dot = &dots[i];
+        // Calculate Maxminum Distance
 
-        if (dot->type != 'L') {
-            continue;
+        if (i != start_index && land_dots[i * 3 + 1] == land_dots[(i - 1) * 3 + 1]) {
+            int min_dist_sq = (int)sqrt(min_dist) + 1 + land_dots[i * 3] - land_dots[(i - 1) * 3];
+            min_dist = min_dist_sq * min_dist_sq;
+        } else {
+            min_dist = INT_MAX;
         }
 
         // Find Nearest Biome Origin Dot
 
-        int min_dist = INT_MAX;
         int origin_index = 0;
-        const int coord[2] = {dot->x, dot->y};
+        const int coord[2] = {land_dots[i * 3], land_dots[i * 3 + 1]};
         query_recursive(origin_tree_root, coord, 0, &origin_index, &min_dist);
 
         // Set Dot Type
 
-        dot->type = dots[origin_index].type;
+        dots[land_dots[i * 3 + 2]].type = dots[origin_index].type;
 
         atomic_fetch_add(&section_progress[4], 1);
 
@@ -815,7 +953,7 @@ int main(int argc, char *argv[]) {
     // Get Inputs
 
     bool auto_mode;
-    char *output_file;
+    char output_file[229];
     int width, height, map_resolution, island_abundance, coastline_smoothing, processes;
     float island_size;
 
@@ -825,7 +963,7 @@ int main(int argc, char *argv[]) {
 
         auto_mode = false;
 
-        output_file = "result.png";
+        strncpy(output_file, "result.png", 11);
 
         // Get Program Version
 
@@ -928,7 +1066,7 @@ int main(int argc, char *argv[]) {
         island_size = atoi(argv[5]) / 10.0;
         coastline_smoothing = atoi(argv[6]);
         processes = atoi(argv[7]);
-        output_file = argv[8];
+        strncpy(output_file, argv[8], 229);
 
     }
 
@@ -1053,24 +1191,6 @@ int main(int argc, char *argv[]) {
 
     section_progress_total[2] = num_reg_dots;
 
-    // Create Origin Dots Array
-
-
-
-    // Create Piece Starts
-
-    int piece_length = num_reg_dots / processes;
-
-    int piece_starts[processes + 1];
-    for (int i = 0; i < processes; i++) {
-        piece_starts[i] = i * piece_length + num_special_dots;
-    }
-    piece_starts[processes] = num_dots;
-    /*
-    used to create x pieces of size num_reg_dots / x, where x = processes
-    last piece may be larger, special dots are skipped
-    */
-
     // Create Land Origin KDTree
 
     const int num_origin_dots = num_special_dots / 2;
@@ -1082,29 +1202,58 @@ int main(int argc, char *argv[]) {
         land_origin_dots[i * 3 + 2] = i;
     }
     Node *origin_tree_root = NULL;
-    origin_tree_root = build_recursive(num_origin_dots, land_origin_dots, 0);
+    origin_tree_root = build_recursive(land_origin_dots, num_origin_dots, 0);
+
+    // Create and Sort Regular Dots
+
+    int *reg_dots = malloc(num_reg_dots * 3 * sizeof(int));
+    for (int i = num_special_dots; i < num_dots; i++) {
+        Dot *dot = &dots[i];
+        const int index = i - num_special_dots;
+        reg_dots[index * 3] = dot->x;
+        reg_dots[index * 3 + 1] = dot->y;
+        reg_dots[index * 3 + 2] = i;
+    }
+
+    // Create Regular Piece Starts
+
+    int reg_piece_length = num_reg_dots / processes;
+    int reg_piece_starts[processes + 1];
+    for (int i = 0; i < processes; i++) {
+        reg_piece_starts[i] = i * reg_piece_length;
+    }
+    reg_piece_starts[processes] = num_reg_dots;
+    /*
+    used to create x pieces of size num_reg_dots / x, where x = processes
+    last piece may be larger, special dots are skipped
+    */
 
     // Run Workers
 
     int fork_pids[processes]; // Create forks list
     for (int i = 0; i < processes; i++) {
+
         fork_pids[i] = fork(); // Create fork
-        if (fork_pids[i] == 0) {
-            // e.g. biogen-worker00
-            set_process_title("worker", i);
-            assign_sections( // Run worker
-                map_resolution, island_size, piece_starts[i], piece_starts[i + 1],
-                origin_tree_root, dots, section_progress
-            );
-            exit(0); // Kill worker
+        if (fork_pids[i] != 0) {
+            continue;
         }
+
+        // e.g. biogen-worker00
+        set_process_title("worker", i);
+        assign_sections( // Run worker
+            map_resolution, island_size, reg_piece_starts[i], reg_piece_starts[i + 1], reg_dots,
+            origin_tree_root, dots, section_progress
+        );
+        exit(0); // Kill worker
+
     }
     for (int i = 0; i < processes; i++) {
         waitpid(fork_pids[i], NULL, 0); // Wait for workers
     }
 
-    // Free Land Origin Tree
+    // Free Regular Dots and Land Origin Tree
 
+    free(reg_dots);
     free_recursive(origin_tree_root);
 
     // Set Section Completion Time
@@ -1148,31 +1297,57 @@ int main(int argc, char *argv[]) {
         Node *land_tree_root = NULL;
         Node *water_tree_root = NULL;
 
-        land_tree_root = build_recursive(num_land_dots, land_dots, 0);
-        water_tree_root = build_recursive(num_water_dots, water_dots, 0);
+        land_tree_root = build_recursive(land_dots, num_land_dots, 0);
+        water_tree_root = build_recursive(water_dots, num_water_dots, 0);
 
-        free(land_dots);
-        free(water_dots);
+        // Sort Land and Water Dots
+
+        quicksort_recursive(land_dots, 0, num_land_dots - 1, width);
+        quicksort_recursive(water_dots, 0, num_water_dots - 1, width);
+
+        // Create Piece Starts for Land and Water Dots
+
+        int land_piece_length = num_land_dots / processes;
+        int land_piece_starts[processes + 1];
+        for (int i = 0; i < processes; i++) {
+            land_piece_starts[i] = i * land_piece_length;
+        }
+        land_piece_starts[processes] = num_land_dots;
+
+        int water_piece_length = num_water_dots / processes;
+        int water_piece_starts[processes + 1];
+        for (int i = 0; i < processes; i++) {
+            water_piece_starts[i] = i * water_piece_length;
+        }
+        water_piece_starts[processes] = num_water_dots;
 
         // Run Workers
 
         for (int i = 0; i < processes; i++) {
+
             fork_pids[i] = fork();
-            if (fork_pids[i] == 0) {
-                set_process_title("worker", i);
-                smooth_coastlines(
-                    coastline_smoothing, piece_starts[i], piece_starts[i + 1],
-                    land_tree_root, water_tree_root,
-                    num_dots, num_special_dots, num_reg_dots, dots, section_progress
-                );
-                exit(0);
+            if (fork_pids[i] != 0) {
+                continue;
             }
+
+            set_process_title("worker", i);
+            smooth_coastlines(
+                coastline_smoothing,
+                land_dots, land_piece_starts[i], land_piece_starts[i + 1], land_tree_root,
+                water_dots, water_piece_starts[i], water_piece_starts[i + 1], water_tree_root,
+                num_dots, num_land_dots, num_water_dots, dots, section_progress
+            );
+            exit(0);
+
         }
         for (int i = 0; i < processes; i++) {
             waitpid(fork_pids[i], NULL, 0);
         }
 
-        // Free Trees
+        // Free Dot Lists and Trees
+
+        free(land_dots);
+        free(water_dots);
 
         free_recursive(land_tree_root);
         free_recursive(water_tree_root);
@@ -1205,53 +1380,72 @@ int main(int argc, char *argv[]) {
         }
     }
 
-    // Create Piece Starts for All Dots
-
-    piece_length = num_dots / processes;
-    for (int i = 0; i < processes; i++) {
-        piece_starts[i] = i * piece_length;
-    }
-    piece_starts[processes] = num_dots;
-
     // Create Water Biomes
     // Adds ice, depth
 
     // Build Land Dots KDTree
 
     int num_land_dots = 0;
+    int num_water_dots = 0;
     int *land_dots = malloc(num_dots * 3 * sizeof(int));
+    int *water_dots = malloc(num_dots * 3 * sizeof(int));
+
     for (int i = 0; i < num_dots; i++) {
+        const Dot *dot = &dots[i];
         if (dots[i].type == 'L') {
-            const Dot *dot = &dots[i];
             land_dots[num_land_dots * 3] = dot->x;
             land_dots[num_land_dots * 3 + 1] = dot->y;
             land_dots[num_land_dots * 3 + 2] = i;
             num_land_dots++;
+        } else {
+            water_dots[num_water_dots * 3] = dot->x;
+            water_dots[num_water_dots * 3 + 1] = dot->y;
+            water_dots[num_water_dots * 3 + 2] = i;
+            num_water_dots++;
         }
     }
+
     Node *land_tree_root = NULL;
-    land_tree_root = build_recursive(num_land_dots, land_dots, 0);
+    land_tree_root = build_recursive(land_dots, num_land_dots, 0);
     free(land_dots);
+
+    // Sort Water Dots
+
+    quicksort_recursive(water_dots, 0, num_water_dots - 1, width);
+
+    // Create Piece Starts Water Dots
+
+    int water_piece_length = num_water_dots / processes;
+    int water_piece_starts[processes + 1];
+    for (int i = 0; i < processes; i++) {
+        water_piece_starts[i] = i * water_piece_length;
+    }
+    water_piece_starts[processes] = num_water_dots;
 
     // Run Workers
 
     for (int i = 0; i < processes; i++) {
+
         fork_pids[i] = fork();
-        if (fork_pids[i] == 0) {
-            set_process_title("worker", i);
-            generate_biomes_water(
-                piece_starts[i], piece_starts[i + 1], height, land_tree_root,
-                num_dots, dots, section_progress
-            );
-            exit(0);
+        if (fork_pids[i] != 0) {
+            continue;
         }
+
+        set_process_title("worker", i);
+        generate_biomes_water(
+            water_piece_starts[i], water_piece_starts[i + 1], water_dots, land_tree_root,
+            height, num_dots, dots, section_progress
+        );
+        exit(0);
+
     }
     for (int i = 0; i < processes; i++) {
         waitpid(fork_pids[i], NULL, 0);
     }
 
-    // Free Land Tree
+    // Free Water Dots and Land Tree
 
+    free(water_dots);
     free_recursive(land_tree_root);
 
     // Add Biome Origin Dots
@@ -1353,28 +1547,59 @@ int main(int argc, char *argv[]) {
         biome_dots[i * 3 + 2] = biome_origin_indexes[i];
     }
     Node *biome_tree_root = NULL;
-    biome_tree_root = build_recursive(num_biome_dots, biome_dots, 0);
+    biome_tree_root = build_recursive(biome_dots, num_biome_dots, 0);
     free(biome_dots);
+
+    // Create and Sort Lands
+    // Original "land_dots" was freed in water biome generation
+
+    int num_land_dots2 = 0;
+    int *land_dots2 = malloc(num_dots * 3 * sizeof(int));
+    for (int i = 0; i < num_dots; i++) {
+        const Dot *dot = &dots[i];
+        if (dot->type == 'L') {
+            land_dots2[num_land_dots2 * 3] = dot->x;
+            land_dots2[num_land_dots2 * 3 + 1] = dot->y;
+            land_dots2[num_land_dots2 * 3 + 2] = i;
+            num_land_dots2++;
+        }
+    }
+
+    quicksort_recursive(land_dots2, 0, num_land_dots2 - 1, width);
+
+    // Create Piece Starts Land Dots
+
+    int land_piece_length = num_land_dots2 / processes;
+    int land_piece_starts[processes + 1];
+    for (int i = 0; i < processes; i++) {
+        land_piece_starts[i] = i * land_piece_length;
+    }
+    land_piece_starts[processes] = num_land_dots2;
 
     // Run Workers
 
     for (int i = 0; i < processes; i++) {
+
         fork_pids[i] = fork();
-        if (fork_pids[i] == 0) {
-            set_process_title("worker", i);
-            generate_biomes_land(
-                piece_starts[i], piece_starts[i + 1], biome_tree_root, num_dots,
-                biome_origin_indexes, dots, section_progress
-            );
-            exit(0);
+        if (fork_pids[i] != 0) {
+            continue;
         }
+
+        set_process_title("worker", i);
+        generate_biomes_land(
+            land_piece_starts[i], land_piece_starts[i + 1], land_dots2,
+            biome_tree_root, biome_origin_indexes, num_dots, dots, section_progress
+        );
+        exit(0);
+
     }
     for (int i = 0; i < processes; i++) {
         waitpid(fork_pids[i], NULL, 0);
     }
 
-    // Free Tree
+    // Free Land Dots and Biome Tree
 
+    free(land_dots2);
     free_recursive(biome_tree_root);
 
     // Set Section Completion Time
@@ -1407,21 +1632,25 @@ int main(int argc, char *argv[]) {
         dot_coords[i * 3 + 2] = i;
     }
     Node *tree_root = NULL;
-    tree_root = build_recursive(num_dots, dot_coords, 0);
+    tree_root = build_recursive(dot_coords, num_dots, 0);
     free(dot_coords);
 
     // Run Workers
 
     for (int i = 0; i < processes; i++) {
+
         fork_pids[i] = fork();
-        if (fork_pids[i] == 0) {
-            set_process_title("worker", i);
-            generate_image(
-                section_starts[i], section_starts[i + 1], width, tree_root,
-                num_dots, dots, image_indexes, type_counts, section_progress
-            );
-            exit(0);
+        if (fork_pids[i] != 0) {
+            continue;
         }
+
+        set_process_title("worker", i);
+        generate_image(
+            section_starts[i], section_starts[i + 1], width, tree_root,
+            num_dots, dots, image_indexes, type_counts, section_progress
+        );
+        exit(0);
+
     }
     for (int i = 0; i < processes; i++) {
         waitpid(fork_pids[i], NULL, 0);
@@ -1439,7 +1668,7 @@ int main(int argc, char *argv[]) {
 
     // --Finish--
 
-    atomic_store(&section_progress_total[6], height);
+    atomic_store(&section_progress_total[6], height * 5);
 
     // Create Image
 
@@ -1558,11 +1787,13 @@ int main(int argc, char *argv[]) {
 
     fclose(fptr);
 
+    atomic_store(&section_progress[6], height * 5);
+
+    // Set Section Completion Time
+
     clock_gettime(CLOCK_REALTIME, &time_now);
     section_times[6] = (float)(time_now.tv_sec - start_time.tv_sec) +
         (time_now.tv_nsec - start_time.tv_nsec) / 1000000000.0 - sum_list_float(section_times, 7);
-
-    // Set Section Completion Time
 
     section_times[7] = (float)(time_now.tv_sec - start_time.tv_sec) +
         (time_now.tv_nsec - start_time.tv_nsec) / 1000000000.0;